other: fix remaining conversion warnings
- Converting datatype to the correct one.

- As macro do not perform type-checking before, so these functions used
  it with different datatype could create a lot of conversion warnings.
  For this reason, these warnings could be consider as normal
  information and will not be fixed.

+++ include/dlt/dlt_common.h
@@ -297,7 +297,7 @@ enum {
 
 #   define DLT_MSG_READ_VALUE(dst, src, length, type) \
     { \
-        if ((length < 0) || ((length) < ((int32_t)sizeof(type)))) \
+        if ((length < 0) || ((length) < ((int32_t) sizeof(type)))) \
         { length = -1; } \
         else \
         { dst = *((type *)src); src += sizeof(type); length -= (int32_t) sizeof(type); } \
@@ -319,7 +319,7 @@ enum {
     } \
     else \
     { \
-        memcpy(dst, src, length); \
+        memcpy(dst, src, (size_t) length); \
         dlt_clean_string(dst, length); \
         dst[length] = 0; \
         src += length; \
